chore: Add pre-commit hook to run linter and code formatter
This commit adds a pre-commit hook to the repository. The hook runs the linter and code formatter before each commit. It checks for modified files before and after running the linter and formatter. If any files are modified, they are added to the list of files to be committed.

The pre-commit hook helps ensure that all code changes adhere to the project's linting and formatting rules.